udpate

+++ sample-blog/app/Http/Controllers/ArticleController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Article;
 use App\Photo;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class ArticleController extends Controller
@@ -13,6 +15,7 @@ class ArticleController extends Controller
     public function __construct()
     {
         $this->middleware("auth");
+//        $this->middleware('isAdmin');
     }
 
     /**
@@ -22,7 +25,11 @@ class ArticleController extends Controller
      */
     public function index()
     {
-        $articles = Article::orderBy("id","desc")->paginate(5);
+        if(Auth::user()->role == 0){
+            $articles = Article::orderBy("id","desc")->paginate(5);
+        }else{
+            $articles = Article::where('user_id',Auth::id())->orderBy("id","desc")->paginate(5);
+        }
         return view("article.index",compact('articles'));
     }
 
@@ -59,7 +66,7 @@ class ArticleController extends Controller
             $fileNameArr = [];
 
             foreach ($request->file("photo") as $file){
-                $newFileName = uniqid()."_profile.".$file->getClientOriginalExtension();
+                $newFileName = uniqid()."_article.".$file->getClientOriginalExtension();
                 array_push($fileNameArr,$newFileName);
                 $dir = "/public/article/";
                 $file->storeAs($dir,$newFileName);
@@ -138,9 +145,23 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
-        $title = $article->title;
-        $article->delete();
-        return redirect()->route("article.index")->with("status","<b>$title</b> is deleted.");
+//        return $article;
+        if(Gate::allows('article-delete',$article)) {
+            if (isset($article->getPhotos)) {
+                $dir = "/public/article/";
+                foreach ($article->getPhotos as $p) {
+                    Storage::delete($dir . $p->location);
+                }
+                $toDel = $article->getPhotos->pluck("id");
+                Photo::destroy($toDel);
+            }
+            $title = $article->title;
+            $article->delete();
+            return redirect()->route("article.index")->with("status", "<b>$title</b> is deleted.");
+        }
+        return abort(404);
+
+
 
     }
